pyproject.toml
[project]
name = "indieclaw"
version = "0.1.0"
description = "Self-hosted Telegram AI agent. 10 tools, skills, subconscious, CLI learning."
readme = "README.md"
requires-python = ">=3.12"
authors = [{name = "IndieClaw Contributors"}]
dependencies = [
    "python-telegram-bot>=20.0",
    "claude-agent-sdk>=0.1.44",
    "apscheduler>=3.10.0",
    "requests>=2.32.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0",
    "loguru>=0.7.0",
    "questionary>=2.0.0",
    "textual>=0.70",
    "typer>=0.24.1",
    "playwright>=1.40.0",
    "edge-tts>=6.1.0",
    "aiohttp>=3.9.0",
]

[project.scripts]
indieclaw = "indieclaw.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["indieclaw", "templates"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.23",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF", "PIE", "C4", "PT", "TCH"]
ignore = [
    "E501",     # line length — handled by formatter when needed
    "SIM105",   # suppressible-exception — contextlib.suppress less readable here
    "UP017",    # datetime.UTC — keep timezone.utc for 3.12 compat with older SDKs
    "B904",     # raise-without-from — often intentional in tool wrappers
    "SIM108",   # ternary — sometimes less readable
    "RUF012",   # mutable class var — Config.DEFAULTS is intentionally mutable at class level
    "TCH",      # type-checking imports — adds complexity for minimal benefit in this codebase
]

[tool.bandit]
skips = ["B404", "B603", "B607", "B108"]

[tool.ruff.lint.isort]
known-first-party = ["indieclaw"]

